+++ packages/web3-core-requestmanager/src/jsonrpc.js
@@ -27,8 +27,9 @@
 // Initialize Jsonrpc as a simple object with utility functions.
 var Jsonrpc = {
     // This is the starting counter for the Jsonrpc.id
-    // Pick a random number between 0 and half of the max safe integer to stay below the max when incrementing
-    messageId: Math.floor(Math.random() * Number.MAX_SAFE_INTEGER / 2)
+    // Pick a random number between 0 and (the maximum safe integer minus the max unsigned 32 bit integer)
+    // Actually, to insure staying in the safe range while incrementing, we insure making the range less than the max unsigned 32 bit integer
+    messageId: Math.floor(Math.random() * (Number.MAX_SAFE_INTEGER - Math.pow( 2, 32 )))
 };
 
 /**