+++ CHANGELOG.md
@@ -592,3 +592,4 @@ Released with 1.0.0-beta.37 code base.
 
 ### Fixed
 -  Browser builds support polyfills (#5031) (#5053) (#4659) (#4767)
+-  Start incrementing jsonrpc.id from random number (#5327)